Adding role checks for map data page
