+++ apps/frontend/package.json
@@ -33,6 +33,7 @@
     "common": "workspace:*",
     "dayjs": "^1.11.10",
     "framer-motion": "^11.0.24",
+    "jwt-decode": "^4.0.0",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
     "postcss": "^8.4.35",
