+++ apps/frontend/src/routes/EmployeeTable.tsx
@@ -232,7 +232,7 @@ const EmployeeTable = () => {
                       </div>
                     </div>
                   </th>
-                  <td className="px-6 py-4">Edger</td>
+                  <td className="px-6 py-4">Auth0 God</td>
                   <td className="px-6 py-4">
                     <div className="flex items-center">
                       <div className="h-2.5 w-2.5 rounded-full bg-green-500 me-2"></div>
