+++ package.json
@@ -34,6 +34,7 @@
   },
   "packageManager": "yarn@4.0.2",
   "devDependencies": {
+    "@types/jsonwebtoken": "^9",
     "@types/material-ui": "^0",
     "@types/multer": "^1.4.11",
     "@types/react": "^18.0.37",
@@ -61,6 +62,7 @@
     "@yarnpkg/pnpify": "^4.0.0-rc.45",
     "dotenv": "^16.4.5",
     "dotenv-cli": "^7.4.0",
+    "jsonwebtoken": "^9.0.2",
     "material-ui": "^0.20.2",
     "multer": "^1.4.5-lts.1",
     "react": "^18.2.0",
